lambda-transcribe/requirements.txt
openai-whisper>=20231117
torch>=2.0.0,<2.5.0
torchaudio>=2.0.0,<2.5.0
transformers>=4.30.0
pyannote.audio>=3.1.0
matplotlib>=3.7.0
requests>=2.31.0
boto3>=1.34.0
ffmpeg-python>=0.2.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
